Cli: Multi input and log support

+++ beet.py
@@ -7,6 +7,7 @@ Cli utility for tracking bees in the beemon project.
 '''
 
 import argparse
+import os
 
 import cv2
 
@@ -15,26 +16,12 @@ from beet.video_policy import VideoPolicy
 
 __author__ = "verbetam"
 
-inputfile = ""
-verbose = False
-mask = False
-visible = False
-remote = False
-auth = False
-tracks = False
-contours = False
-boundary = False
-
 
 def parse_args(parser):
-    global inputfile, verbose, mask, visible, remote, auth, tracks,\
-            boundary, contours
-
     parser.add_argument(
-        '-i', '--input',
-        help="Input file path",
-        dest="input",
-        required=True)
+        'files',
+        metavar='FILE',
+        nargs='*')
     parser.add_argument(
         '-V', '--verbose',
         help="Produce verbose output",
@@ -53,6 +40,13 @@ def parse_args(parser):
         action="store_true",
         dest="visible",
         default=False)
+    parser.add_argument(
+        '-l', '--log',
+        type=argparse.FileType('w'),
+        help="Specifies where where to log output to",
+        dest="log",
+        default='beet.log'
+    )
     parser.add_argument(
         '-r', '--remote',
         help="Access source files through FTP",
@@ -82,36 +76,55 @@ def parse_args(parser):
         help="Access source files via FTP with info in given file",
         default=False)
     args = parser.parse_args()
-
-    inputfile = args.input
-    verbose = args.verbose
-    mask = args.mask
-    visible = args.visible
-    remote = args.remote
-    auth = args.auth
-    tracks = args.tracks
-    contours = args.contours
-    boundary = args.boundary
     return args
 
 
 def main():
     parser = argparse.ArgumentParser(description='Bee tracker utility')
     args = parse_args(parser)
+    if (args.files is None and not args.remote):
+        print("\nPlease provide an input file(s) or specifiy remote (-r)\n")
+        return
     video_policy = VideoPolicy(args)
     source = video_policy.get_video()
 
-    app = beet.kalman_track.App(invisible=visible,
-                                draw_contours=contours,
-                                draw_tracks=tracks,
-                                draw_boundary=boundary)
-    app.openNewVideo(source)
-    app.run()
+    # Process each file given.
+    for filepath in source:
+        # If the file is valid.
+        if os.path.exists(filepath):
+            # Create a new app with given options.
+            app = beet.kalman_track.App(invisible=not args.visible,
+                                        draw_contours=args.contours,
+                                        draw_tracks=args.tracks,
+                                        draw_boundary=args.boundary)
+            # Open the current file.
+            app.openNewVideo(filepath)
+            # Run the app.
+            app.run()
 
-    if(verbose):
-        print("Arrivals: {0} Departures: {1}".format(app.arrivals,
-                                                     app.departures))
-    cv2.destroyAllWindows()
+            # If verbose, print results.
+            if(args.verbose):
+                line = "File: {0}\n  Arrivales: {1}\n  Departures: {2}\n"
+                print(line.format(filepath, app.arrivals,
+                                  app.departures, app.departures))
+
+            # If logging, log results.
+            if(args.log):
+                line = "File: {0}\n  Arrivales: {1}\n  Departures: {2}\n"
+                args.log.write(line.format(filepath, app.arrivals,
+                                           app.departures))
+
+            cv2.destroyAllWindows()
+
+        # If file is invalid.
+        else:
+            line = "Error: {0} is not a valid file path\n".format(filepath)
+            # Print error.
+            print(line)
+
+            # If logging, log error.
+            if(args.log):
+                args.log.write(line)
 
 
 if __name__ == "__main__":
